+++ src/main/java/org/jabref/gui/entryeditor/EntryEditor.java
@@ -33,7 +33,6 @@ import org.jabref.gui.undo.UndoableKeyChange;
 import org.jabref.gui.util.DefaultTaskExecutor;
 import org.jabref.logic.TypedBibEntry;
 import org.jabref.logic.bibtexkeypattern.BibtexKeyGenerator;
-import org.jabref.logic.help.HelpFile;
 import org.jabref.logic.importer.EntryBasedFetcher;
 import org.jabref.logic.importer.WebFetchers;
 import org.jabref.logic.search.SearchQueryHighlightListener;
@@ -68,10 +67,11 @@ public class EntryEditor extends BorderPane {
      * A reference to the entry this editor works on.
      */
     private BibEntry entry;
+    private SourceTab sourceTab;
+
     @FXML private TabPane tabbed;
     @FXML private Button typeChangeButton;
     @FXML private Button fetcherButton;
-    private SourceTab sourceTab;
     @FXML private Label typeLabel;
     private final EntryEditorPreferences preferences;
     private final DialogService dialogService;
@@ -277,7 +277,6 @@ public class EntryEditor extends BorderPane {
         ContextMenu typeMenu = new ChangeEntryTypeMenu(preferences.getKeyBindings()).getChangeEntryTypePopupMenu(entry, bibDatabaseContext, undoManager);
         typeLabel.setOnMouseClicked(event -> typeMenu.show(typeLabel, Side.RIGHT, 0, 0));
         typeChangeButton.setOnMouseClicked(event -> typeMenu.show(typeChangeButton, Side.RIGHT, 0, 0));
-
         // Add menu for fetching bibliographic information
         ContextMenu fetcherMenu = new ContextMenu();
         for (EntryBasedFetcher fetcher : WebFetchers.getEntryBasedFetchers(preferences.getImportFormatPreferences())) {
@@ -305,9 +304,4 @@ public class EntryEditor extends BorderPane {
             }
         });
     }
-
-    private String convertToHex(java.awt.Color color) {
-        return String.format("#%02x%02x%02x", color.getRed(), color.getGreen(), color.getBlue());
-    }
-
 }