+++ src/main/java/org/jabref/logic/importer/fetcher/IEEE.java
@@ -18,32 +18,43 @@ import org.slf4j.LoggerFactory;
 
 /**
  * Class for finding PDF URLs for entries on IEEE
- * Will first look for URLs of the type http://ieeexplore.ieee.org/stamp/stamp.jsp?[tp=&]arnumber=...
+ * Will first look for URLs of the type https://ieeexplore.ieee.org/stamp/stamp.jsp?[tp=&]arnumber=...
  * If not found, will resolve the DOI, if it starts with 10.1109, and try to find a similar link on the HTML page
  */
 public class IEEE implements FulltextFetcher {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(IEEE.class);
+    private static final String STAMP_BASE_STRING_DOCUMENT = "/stamp/stamp.jsp?tp=&arnumber=";
     private static final Pattern STAMP_PATTERN = Pattern.compile("(/stamp/stamp.jsp\\?t?p?=?&?arnumber=[0-9]+)");
-    private static final Pattern PDF_PATTERN = Pattern
-            .compile("\"(http://ieeexplore.ieee.org/ielx[0-9/]+\\.pdf[^\"]+)\"");
+    private static final Pattern DOCUMENT_PATTERN = Pattern.compile("document/([0-9]+)/");
+
+    private static final Pattern PDF_PATTERN = Pattern.compile("\"(https://ieeexplore.ieee.org/ielx[0-9/]+\\.pdf[^\"]+)\"");
     private static final String IEEE_DOI = "10.1109";
-    private static final String BASE_URL = "http://ieeexplore.ieee.org";
+    private static final String BASE_URL = "https://ieeexplore.ieee.org";
 
     @Override
     public Optional<URL> findFullText(BibEntry entry) throws IOException {
         Objects.requireNonNull(entry);
 
         String stampString = "";
+
         // Try URL first -- will primarily work for entries from the old IEEE search
         Optional<String> urlString = entry.getField(FieldName.URL);
         if (urlString.isPresent()) {
-            // Is the URL a direct link to IEEE?
-            Matcher matcher = STAMP_PATTERN.matcher(urlString.get());
-            if (matcher.find()) {
-                // Found it
-                stampString = matcher.group(1);
+
+            Matcher documentUrlMatcher = DOCUMENT_PATTERN.matcher(urlString.get());
+            if (documentUrlMatcher.find()) {
+                String docId = documentUrlMatcher.group(1);
+                stampString = STAMP_BASE_STRING_DOCUMENT + docId;
             }
+
+            //You get this url if you export bibtex from IEEE
+            Matcher stampMatcher = STAMP_PATTERN.matcher(urlString.get());
+            if (stampMatcher.find()) {
+                // Found it
+                stampString = stampMatcher.group(1);
+            }
+
         }
 
         // If not, try DOI
@@ -51,7 +62,11 @@ public class IEEE implements FulltextFetcher {
             Optional<DOI> doi = entry.getField(FieldName.DOI).flatMap(DOI::parse);
             if (doi.isPresent() && doi.get().getDOI().startsWith(IEEE_DOI) && doi.get().getExternalURI().isPresent()) {
                 // Download the HTML page from IEEE
-                String resolvedDOIPage = new URLDownload(doi.get().getExternalURI().get().toURL()).asString();
+                URLDownload urlDownload = new URLDownload(doi.get().getExternalURI().get().toURL());
+                //We don't need to modify the cookies, but we need support for them
+                urlDownload.getCookieFromUrl();
+
+                String resolvedDOIPage = urlDownload.asString();
                 // Try to find the link
                 Matcher matcher = STAMP_PATTERN.matcher(resolvedDOIPage);
                 if (matcher.find()) {
@@ -67,7 +82,11 @@ public class IEEE implements FulltextFetcher {
         }
 
         // Download the HTML page containing a frame with the PDF
-        String framePage = new URLDownload(BASE_URL + stampString).asString();
+        URLDownload urlDownload = new URLDownload(BASE_URL + stampString);
+        //We don't need to modify the cookies, but we need support for them
+        urlDownload.getCookieFromUrl();
+
+        String framePage = urlDownload.asString();
         // Try to find the direct PDF link
         Matcher matcher = PDF_PATTERN.matcher(framePage);
         if (matcher.find()) {