+++ src/main/java/org/jabref/logic/importer/fetcher/LibraryOfCongress.java
@@ -6,6 +6,7 @@ import java.net.URL;
 
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.importer.IdBasedParserFetcher;
+import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.Parser;
 import org.jabref.logic.importer.fileformat.ModsImporter;
 
@@ -15,6 +16,12 @@ import org.apache.http.client.utils.URIBuilder;
  * Fetcher for the Library of Congress Control Number (LCCN) using https://lccn.loc.gov/
  */
 public class LibraryOfCongress implements IdBasedParserFetcher {
+
+    private final ImportFormatPreferences importFormatPreferences;
+
+    public LibraryOfCongress(ImportFormatPreferences importFormatPreferences) {
+        this.importFormatPreferences = importFormatPreferences;
+    }
 
     @Override
     public String getName() {
@@ -29,6 +36,6 @@ public class LibraryOfCongress implements IdBasedParserFetcher {
 
     @Override
     public Parser getParser() {
-        return new ModsImporter();
+        return new ModsImporter(this.importFormatPreferences);
     }
 }